Reorganize storage perf utils
