+++ imagetest/test_suites/storageperf/storage_perf_utils.go
@@ -1,12 +1,8 @@
 package storageperf
 
 import (
-	"encoding/json"
 	"fmt"
 	"os/exec"
-	"path/filepath"
-	"strconv"
-	"strings"
 
 	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest/utils"
 )
@@ -47,46 +43,6 @@ type FIOStatistics struct {
 	X         map[string]interface{} `json:"-"`
 }
 
-// BlockDeviceList gives full information about blockdevices, from the output of lsblk.
-type BlockDeviceList struct {
-	BlockDevices []BlockDevice `json:"blockdevices,omitempty"`
-}
-
-// BlockDevice defines information about a single partition or disk in the output of lsblk.
-type BlockDevice struct {
-	Name string `json:"name,omitempty"`
-	Size string `json:"size,omitempty"`
-	Type string `json:"type,omitempty"`
-	// Other fields are not currently used.
-	X map[string]interface{} `json:"-"`
-}
-
-// This method currently only runs the commands for linux. Support for getting the partition on windows will be added in the future.
-func getMountDiskPartition(diskExpectedSizeGb int) (string, error) {
-	lsblkCmd := "lsblk"
-	if !utils.CheckLinuxCmdExists(lsblkCmd) {
-		return "", fmt.Errorf("could not find lsblk")
-	}
-	lsblkout, err := exec.Command(lsblkCmd, "-o", "name,size,type", "--json").CombinedOutput()
-	if err != nil {
-		return "", fmt.Errorf("failed to execute lsblk cmd with error: %v", err)
-	}
-
-	var blockDevices BlockDeviceList
-	if err := json.Unmarshal(lsblkout, &blockDevices); err != nil {
-		return "", fmt.Errorf("failed to unmarshal lsblk output with error: %v", err)
-	}
-
-	diskExpectedSizeGbString := strconv.Itoa(diskExpectedSizeGb) + "G"
-	for _, blockDev := range blockDevices.BlockDevices {
-		if strings.ToLower(blockDev.Type) == "disk" && blockDev.Size == diskExpectedSizeGbString {
-			return blockDev.Name, nil
-		}
-	}
-
-	return "", fmt.Errorf("disk block with size not found")
-}
-
 // installFioWindows copies the fio.exe file onto the VM instance.
 func installFioWindows() error {
 	if procStatus, err := utils.RunPowershellCmd("gsutil cp " + fioWindowsGCS + " " + fioWindowsLocalPath); err != nil {
@@ -118,13 +74,4 @@ func installFioLinux() error {
 		return fmt.Errorf("install fio command failed with errors: %v", err)
 	}
 	return nil
-}
-
-func getMountDiskPartitionSymlink() (string, error) {
-	mountDiskSymlink := "/dev/disk/by-id/google-" + mountDiskName
-	symlinkRealPath, err := filepath.EvalSymlinks(mountDiskSymlink)
-	if err != nil {
-		return "", fmt.Errorf("symlink could not be resolved with error: %v", err)
-	}
-	return symlinkRealPath, nil
 }
